Fixed unnessary warnings with VS2017

+++ neo/sys/win32/win_main.cpp
@@ -878,7 +878,7 @@ Sys_DLL_Load
 intptr_t Sys_DLL_Load( const char *dllName )
 {
 	HINSTANCE libHandle = LoadLibrary( dllName );
-	return (int)libHandle;
+	return (intptr_t)libHandle;
 }
 
 /*
